Fix DocumentNode.find()
